Inherit COOP: restrict-properties and un-suppress opener
To make the `Cross-Origin-Opener-Policy: restrict-properties` (COOP:RP)
more useful, we'd like to let cross-origin iframes inherit this policy
from their parent and keep the opener connection between windows that
the iframes open. See
https://github.com/hemeryar/coi-with-popups/blob/main/docs/cross_origin_iframe_popup.MD
for more explanations.

In https://crrev.com/c/4404116, we've bundled the origin that sets the
COOP with the policy.  This CL is to use the stored origin in COOP
algorithms:
1. In `RenderFrameHostImpl::DidCommitNewDocument`, we let embedded
documents inherit the COOP:RP bundle when the policy is COOP:RP.
2. To keep the opener connection, we return false in
`COOPSuppressOpener` when COOP value is RP.
3. Since we allow the popup's initial empty document to stay in the
process, we had to restrict its crossOriginIsolated capability.
Otherwise, a cross-origin iframe can get around permission policy that
restricts it from using crossOriginIsolated APIs. Therefore, we pass
down a boolean from browser to renderer saying whether initial empty
documents or their subdocuments can be crossOriginIsolated. We store
this boolean at the top frame and add a method for subframes to query
every time before computing crossOriginIsolated.

Bug: 1385827
Change-Id: I39cd140f70b628114205640af2220ec2de45ac06