Shared Storage: Add sharedStorageWritable option to fetch
We add the option to mark a request as eligible for writing to Shared
Storage, i.e. we add `sharedStorageWritable` to the `RequestInit` IDL
dictionary for `fetch()` and plumb it through to
`network::ResourceRequest`, subject to a `PermissionsPolicy` that may
change the value of the bit for redirects.

Tests will be added in the follow-up CL https://crrev.com/c/4519162.

This is part of a project to add support to writing to shared storage
from response headers. See
https://github.com/WICG/shared-storage/pull/70 for more information
and https://crrev.com/c/4444068 for a working prototype for `fetch()`.

Bug:1434529,1218540
Change-Id: Ib133492bd90e8cc37b9bc0c7bc61861e0b2afc63